+++ pkgs/artisanal/lib/plugins.dart
@@ -36,3 +36,4 @@ export 'src/plugins/remote_surface_protocol.dart'
 export 'src/plugins/remote_surface_transport.dart'
     show RemotePluginJsonTransport;
 export 'src/plugins/remote_surface_channel.dart' show RemotePluginJsonChannel;
+export 'src/plugins/remote_surface_process.dart' show RemotePluginProcess;